* Warning about signed/unsigned comparison fixed
svn path=/trunk/flamerobin/; revision=984

+++ src/Isaac.h
@@ -137,7 +137,7 @@ public:
         ub4 pw[33];
         for (int i=0; i<33; ++i)
             pw[i] = 0;
-        for (int i=0; i<pwd.Length() && i<32; ++i)
+        for (int i=0; i < (int) pwd.Length() && i<32; ++i)
             pw[i] = (ub4)pwd[i];
 
         for (int j=0; j<32; ++j)
